Moved `QualityExtraction` to core\syntax.py
